Move Python snapshot logic to its own file
Nonfunctional change.
